% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linear_interpolation_cpp}
\alias{linear_interpolation_cpp}
\title{Main function of mapprox() in R}
\usage{
linear_interpolation_cpp(x_r, y, xout_r, rule, verbose)
}
\arguments{
\item{y}{Reference data of a target variable.
This must be a numeric vector with length \code{nrow(x)}.}

\item{rule}{A scalar integer determining how interpolation outside the
given explanatory variables works. This must be eigher
\code{1}, \code{2}, or \code{3}.
If \code{rule = 1}, \code{NA}s are returned for such points.
If \code{rule = 2}, the values at the closest points in the region
given as the reference data are returned.
If \code{rule = 3}, linear extension of the interpolated result is applied.}

\item{verbose}{Logical. Should progress of the process printed?}

\item{x}{Reference data of explanatory variables.
This must be a data frame in which elements are all numeric.
Also, this should be ordered by \code{order} function.}

\item{xout}{Values of explanatory variables where interpolation is to
take place. This must be a data frame and
\code{ncol(xout)} must be the same as \code{ncol(x)}.}
}
\value{
A numeric vector of interpolated result.
}
\description{
\code{mapprox} is an wrapper of \code{linear_interpolation} and
\code{linear_interpolation_cpp}.
}
\details{
This is the main function of \code{mapprox(..., use_cpp = TRUE)}.
Provided data are cleaned beforehand in \code{mapprox} and
then passed to this function.
\code{x} must be completely gridded data,
and duplicated values are not allowed.
Also, no NA in \code{x}, \code{y}, and \code{xout} is allowed.
}
\seealso{
[mapprox()]
}
