% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapprox.R
\name{mapprox}
\alias{mapprox}
\title{Linear interpolation of a multivariable function.}
\usage{
mapprox(x, y, xout, rule = 1)
}
\arguments{
\item{x}{Reference data of explanatory variables.
This must be a vector, matrix or data frame in which elements are all numeric.}

\item{y}{Reference data of a target variable.
This must be a numeric vector with length \code{nrow(x)}.}

\item{xout}{Values of explanatory variables where interpolation is to
take place. This must be a matrix or data frame and
\code{ncol(xout)} must be the same as \code{ncol(x)}.}

\item{rule}{A scalar integer determining how interpolation outside the
given explanatory variables works. This must be eigher
\code{1}, \code{2}, or \code{3}.
If \code{rule = 1}, \code{NA}s are returned for such points.
If \code{rule = 2}, the values at the closest points in the region
given as the reference data are returned.
If \code{rule = 3}, linear extension of the interpolated result is applied.}
}
\value{
A list with components \code{x} and \code{y}.
  The element \code{x} is identical to input \code{xout},
  if its all rows are allowed as appropriate input.
  The element \code{y} is a numeric vector with length \code{nrow(xout)},
  containing interpolated results.
}
\description{
\code{mapprox} is an extension of \code{approx} to deal with multivariable x.
}
\examples{
# An example of one variable: comparison with approx().
x <- 1:10
y <- c(1:3, 8:10, 7:4)
xout <- seq(0, 11, 0.2)
res_approx <- approx(x, y, xout, method = 'linear', rule = 2)
res_mapprox <- mapprox(x, y, xout, rule = 2)
plot(res_approx, pch = 16, cex = 0.5)
points(res_mapprox)
points(x, y, pch = 1, cex = 2)
# legend()


}
\seealso{
[approx()]
}
